Add support for Iedit.
It has the following faces :
- `iedit-occurrence` which is the "normal" face, just press C-; to start iedit on a point, and you should see it
- `iedit-read-only-occurrence` that one is a special face, much rarer in day to day use.

The following note explains how to get `iedit-read-only-occurrence` to show, in case we want to take a screenshot of the mode :
Basically, let's say you do C-; on "foo".
Now, all of the instances of "foo", in the region / buffer you are working with, are selected by iedit.
Let's now say that one of these "foo" is in read-only, and another isn't (meaning that some or all characters of one "foo" can't be modified, while all of the characters of another "foo" can be modified), then, the "foo" with read-only parts, will be shown in this face.
Note that if the entire region/buffer is in read only, then the face used is `iedit-occurrence`.
Basically, the goal of this face is to alert the user that, if they want to modify all instances of "foo", they may run into an issue, because, modifications to the not-readonlys "foo" don't modify the read-only "foo"; that is, even if the "foo" with read-only isn't totally read-only.
It means that, even if only the "f" is read-only, changing "foo" to "fooo" on the non-readonly won't change the read-only to "foo". The converse is also true.